Move sorting to NewPackage.
