+++ nomdl/codegen/codegen_test.go
@@ -196,7 +196,7 @@ func TestGenerateDeps(t *testing.T) {
 	depender := types.NewPackage([]types.Type{}, []ref.Ref{leaf1Ref})
 	dependerRef := types.WriteValue(depender, cs)
 
-	top := types.NewPackage([]types.Type{}, []ref.Ref{dependerRef, leaf2Ref})
+	top := types.NewPackage([]types.Type{}, []ref.Ref{leaf2Ref, dependerRef})
 	types.RegisterPackage(&top)
 
 	localPkgs := refSet{top.Ref(): true}
